+++ src/apps/koyo/aurora/koyo.pool.token-fetcher.ts
@@ -29,6 +29,7 @@ export class AuroraKoyoPoolTokenFetcher implements PositionFetcher<AppTokenPosit
       vaultAddress: '0x0613adbd846cb73e65aa474b785f52697af04c0b',
       resolvePoolTokenAddresses: this.theGraphPoolTokenDataStrategy.build({
         subgraphUrl: 'https://api.thegraph.com/subgraphs/name/koyo-finance/exchange-subgraph-aurora',
+        skipVolume: true,
       }),
     });
   }