+++ src/apps/plutus/arbitrum/plutus.lock.contract-position-fetcher.ts
@@ -1,6 +1,7 @@
 import { Inject } from '@nestjs/common';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
+import { ZERO_ADDRESS } from '~app-toolkit/constants/address';
 import { Register } from '~app-toolkit/decorators';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { ContractPosition } from '~position/position.interface';
@@ -15,6 +16,12 @@ const appId = PLUTUS_DEFINITION.id;
 const groupId = PLUTUS_DEFINITION.groups.lock.id;
 const network = Network.ARBITRUM_MAINNET;
 
+const labels = {
+  [VAULTS.PLS_LOCK_1_MONTH]: 'PLS 1 Month Lock',
+  [VAULTS.PLS_LOCK_3_MONTH]: 'PLS 3 Month Lock',
+  [VAULTS.PLS_LOCK_6_MONTH]: 'PLS 6 Month Lock',
+};
+
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class ArbitrumPlutusLockContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
@@ -28,13 +35,35 @@ export class ArbitrumPlutusLockContractPositionFetcher implements PositionFetche
         network,
         appId,
         groupId,
-        resolveFarmAddresses: () => [VAULTS.PLS_LOCK],
+        dependencies: [
+          {
+            appId: PLUTUS_DEFINITION.id,
+            groupIds: [PLUTUS_DEFINITION.groups.plsDpx.id, PLUTUS_DEFINITION.groups.plsJones.id],
+            network,
+          },
+        ],
+        resolveLabel: (address: string) => labels[address],
+        resolveFarmAddresses: () => [VAULTS.PLS_LOCK_1_MONTH, VAULTS.PLS_LOCK_3_MONTH, VAULTS.PLS_LOCK_6_MONTH],
         resolveFarmContract: ({ address, network }) =>
           this.plutusContractFactory.plutusEpochStaking({ address, network }),
         resolveStakedTokenAddress: ({ contract, multicall }) => multicall.wrap(contract).pls(),
-        resolveRewardTokenAddresses: ({ contract, multicall }) => multicall.wrap(contract).stakingRewards(),
+        resolveRewardTokenAddresses: async ({ contract, multicall }) => {
+          const stakingRewardsAddress = await multicall.wrap(contract).stakingRewards();
+          if (stakingRewardsAddress === ZERO_ADDRESS) return [];
+
+          const stakingRewardsContract = this.plutusContractFactory.plutusEpochStakingRewardsRolling({
+            address: stakingRewardsAddress,
+            network,
+          });
+
+          return Promise.all([
+            multicall.wrap(stakingRewardsContract).plsDPX(),
+            multicall.wrap(stakingRewardsContract).plsJONES(),
+          ]);
+        },
         resolveRois: async () => ({ dailyROI: 0, weeklyROI: 0, yearlyROI: 0 }),
       });
+
     return positions;
   }
 }